trifirst/requirements.txt
fastapi
uvicorn
python-dotenv
requests
sqlite3  # built-in
httpx
